requirements.txt
langchain_openai==0.1.16
langchain_experimental==0.0.62
langchainhub==0.1.20
pinecone-client==4.1.2
PyMuPDF==1.24.7
langchain==0.2.7
wikipedia==1.4.0
openai==1.35.13
numpy==1.26.4
streamlit==1.36.0
numexpr==2.10.1
pandas-datareader==0.10.0
duckduckgo_search==6.1.12
scikit-learn==1.5.1
matplotlib==3.9.1
ipywidgets==8.1.3
tqdm==4.66.4
ipykernel==6.29.5
PyYAML==6.0.1
langchain-google-genai==1.0.7
httpx==0.27.2